Extract gallery upload-after helpers to shared test module
