+++ test/browser/tests/attachments/gallery.test.js
@@ -1,6 +1,7 @@
 import { test } from "../../test_helper.js"
 import { expect } from "@playwright/test"
 import { mockActiveStorageUploads } from "../../helpers/active_storage_mock.js"
+import { uploadStandaloneAfter } from "../../helpers/gallery_test_helpers.js"
 
 test.describe("Gallery", () => {
   test.beforeEach(async ({ page }) => {
@@ -402,54 +403,6 @@ async function selectGalleryAtOffset(page, editor, offset, galleryIndex = 0) {
     await selectGalleryImage(page, offset - 1, galleryIndex)
     await editor.send("ArrowRight")
   }
-}
-
-async function uploadStandaloneAfter(editor, anchorType, filePath) {
-  await positionCursorAfterNode(editor, anchorType)
-  await editor.send("x", "Enter")
-  await editor.uploadFile(filePath)
-  await expect(editor.content.locator("figure.attachment--preview > progress")).toHaveCount(0)
-  await editor.flush()
-  await removeBufferParagraphsBetweenImages(editor)
-}
-
-async function positionCursorAfterNode(editor, anchorType) {
-  await editor.locator.evaluate((el, type) => {
-    return new Promise((resolve) => {
-      el.editor.update(() => {
-        const root = el.editor.getEditorState()._nodeMap.get("root")
-        for (const child of root.getChildren()) {
-          if (child.getType() === type) {
-            const next = child.getNextSibling()
-            if (next?.getType() === "provisonal_paragraph") {
-              next.selectStart()
-            } else {
-              child.selectNext(0, 0)
-            }
-            return
-          }
-        }
-      }, { onUpdate: resolve })
-    })
-  }, anchorType)
-}
-
-async function removeBufferParagraphsBetweenImages(editor) {
-  await editor.locator.evaluate((el) => {
-    return new Promise((resolve) => {
-      el.editor.update(() => {
-        const root = el.editor.getEditorState()._nodeMap.get("root")
-        const isImageNode = (n) =>
-          n?.getType() === "action_text_attachment" || n?.getType() === "image_gallery"
-        for (const node of root.getChildren()) {
-          const type = node.getType()
-          if (type !== "paragraph" && type !== "provisonal_paragraph") continue
-          if (!isImageNode(node.getPreviousSibling()) || !isImageNode(node.getNextSibling())) continue
-          if (node.getTextContent().replace(/x/g, "") === "") node.remove()
-        }
-      }, { onUpdate: resolve })
-    })
-  })
 }
 
 async function selectGalleryNode(editor, galleryIndex = 0) {
